src/python-server/requirements.txt
# Core server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Text-to-Speech (Offline Neural Voices)
piper-tts>=1.2.0

# IPA Transcription
gruut>=2.3.0

# PDF Processing
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
